# Conflicts:
#	GritAcademyAPI/src/main/java/com/example/GritAcademyAPI/studentsCourses/StudentsCourses.java

+++ GritAcademyAPI/src/main/java/com/example/GritAcademyAPI/students/StudentsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface StudentsRepository extends JpaRepository<Students, Integer> {
+public interface StudentsRepository extends JpaRepository<Students, Long> {
 
     //till för att hantera om det är förnamn, efternamn eller stad som söks
     List<Students> findByFnameOrLnameOrTown(String fname, String lname, String town);
